# Conflicts:
#	Assets/Scripts/MiuraScript/Gimic.meta
#	Assets/Scripts/MiuraScript/Object/Wall/Wall.cs.meta
#	Assets/Scripts/motai/class.meta
#	Assets/Scripts/motai/class/PlayerClass.cs.meta

+++ Assets/Scripts/Suzumura/EnemyData.cs
@@ -14,7 +14,7 @@ public class EnemyData
     public int TargetPosY;              //
     public int TargetPosZ;              //
     public Sprite sprite;               // 画像読み込み(パス)
-    public int AppearanceTime;          // 出現時間
+    public float AppearanceTime;          // 出現時間
     public int State;                   // 状態
 }
